+++ lib/entities/life_item.dart
