+++ lib/models/common/life_event_params/target_life_item_params.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/foundation.dart';
 import 'package:freezed_annotation/freezed_annotation.dart';
 
-import '../life_item.dart';
+import '../../../entities/life_item.dart';
 
 part 'target_life_item_params.freezed.dart';
 part 'target_life_item_params.g.dart';
