+++ test/unit/models/life_stage_model_test.dart
@@ -1,5 +1,5 @@
 import 'package:HumanLifeGame/api/firestore/store.dart';
-import 'package:HumanLifeGame/models/common/life_item.dart';
+import 'package:HumanLifeGame/entities/life_item.dart';
 import 'package:HumanLifeGame/models/play_room/life_stage.dart';
 import 'package:cloud_firestore_mocks/cloud_firestore_mocks.dart';
 import 'package:flutter_test/flutter_test.dart';
